1.01
Designed for use with [The J A Mortram WordPress
Theme](http://wordpress.org/themes/the-j-a-mortram).

Once activated, plugin adds 'Share This Story' buttons to the end of a
single post before the comments. Posts can be shared on Twitter,
Facebook and Google+. This plugin does not embed tags from any of the
above services in your WordPress site. Instead it simply posts the page
URL and any other information relevant to each network's API.